style: 폴더 구조 수정
- 댓글 기능(feature) 분리
