+++ src/features/comment/components/comment-item/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentItem';
