+++ src/features/comment/components/comment-list/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentList';
