+++ src/features/comment/components/comments/index.ts
@@ -1,0 +1,1 @@
+export * from './Comments';
