+++ src/features/comment/components/create-comment-modal-content/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateCommentModalContent';
