+++ src/features/comment/components/index.ts
@@ -1,0 +1,4 @@
+export * from './comment-item';
+export * from './comment-list';
+export * from './comments';
+export * from './create-comment-modal-content';
