+++ src/features/comment/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useComments';
+export * from './useCreateComment';
