+++ src/features/comment/hooks/useCreateComment.tsx
@@ -1,7 +1,7 @@
 import { COMMENT_QUERY_KEY } from '@/constants/query-key';
+import { CreateComment } from '@/features/comment/schemas';
 import { getQueryClient } from '@/libs/tanstack-query';
 import { useMutation } from '@tanstack/react-query';
-import { CreateComment } from '../schemas';
 import { createComment } from '../services';
 
 export function useCreateComment(projectId: number) {
