+++ src/features/project/hooks/index.ts
@@ -1,6 +1,4 @@
 export * from './useCheckProjectExists';
-export * from './useComments';
-export * from './useCreateComment';
 export * from './useCreateProject';
 export * from './useDeleteProject';
 export * from './useEditProject';
