+++ src/features/project/schemas/index.ts
@@ -1,4 +1,3 @@
-export * from './comment';
 export * from './new-project';
 export * from './project';
 export * from './team';
