+++ src/features/project/services/index.ts
@@ -1,4 +1,4 @@
-import { CreateComment, NewProject } from '../schemas';
+import { NewProject } from '../schemas';
 
 const BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
 
@@ -308,66 +308,5 @@ export const receivePumati = async ({
     throw error instanceof Error
       ? error
       : new Error('An unexpected error occurred while receiving pumati');
-  }
-};
-
-export const createComment = async (
-  projectId: number,
-  commentData: CreateComment,
-  token: string,
-) => {
-  try {
-    const response = await fetch(
-      `${BASE_URL}/api/projects/${projectId}/comments`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}`,
-        },
-        body: JSON.stringify(commentData),
-      },
-    );
-
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-
-    const data = await response.json();
-
-    return data.data;
-  } catch (error) {
-    console.error('Failed to create comment:', error);
-
-    throw error instanceof Error
-      ? error
-      : new Error('An unexpected error occurred while creating a comment');
-  }
-};
-
-export const getComments = async (
-  projectId: number,
-  cursorTime: string = new Date().toISOString(),
-  cursorId: number = 0,
-  pageSize: number = 10,
-) => {
-  try {
-    const response = await fetch(
-      `${BASE_URL}/api/projects/${projectId}/comments?cursor-id=${cursorId}&cursor-time=${cursorTime}&page-size=${pageSize}`,
-    );
-
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-
-    const data = await response.json();
-
-    return data;
-  } catch (error) {
-    console.error('Failed to get comments:', error);
-
-    throw error instanceof Error
-      ? error
-      : new Error('An unexpected error occurred while getting comments');
   }
 };
